Scanner/requirements.txt
alabaster==0.7.12
alembic==1.6.2
api==0.0.7
APScheduler==3.7.0
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorlover==0.3.0
config==0.5.0.post0
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
Django==3.2.2
django-google-charts==0.1.1
django-scheduler==0.9.3
docutils==0.16
entrypoints==0.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-UUID==0.2
fpdf==1.7.2
greenlet==1.1.0
gunicorn==20.1.0
icalendar==4.0.7
idna==2.9
imagesize==1.2.0
importlib-metadata==4.0.1
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.8.0
nose==1.3.7
notebook==6.3.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-nmap==0.6.4
python3-nmap==1.5.0
pytz==2021.1
pyzmq==22.0.3
redis==3.5.3
requests==2.25.1
retrying==1.3.3
rq==1.8.0
scapy==2.4.5
schedule==1.1.0
Send2Trash==1.5.0
sentry-sdk==1.1.0
simplejson==3.17.2
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.0.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.14
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
Tempita==0.5.2
terminado==0.9.4
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
vulners==1.5.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
zipp==3.4.1
